Refactored filter_attributes() from query_filter.py to util.py
